fontgen-ai/requirements.txt
# AI フォント生成システム - 依存関係
# Python 3.10+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# フォント処理
fonttools>=4.40.0
freetype-py>=2.4.0
Pillow>=10.0.0

# 数値計算・科学計算
numpy>=1.24.0
scipy>=1.10.0

# 画像処理
opencv-python>=4.8.0
scikit-image>=0.21.0

# データ処理
pandas>=2.0.0

# 可視化
matplotlib>=3.7.0
seaborn>=0.12.0

# 設定管理
pyyaml>=6.0
omegaconf>=2.3.0

# プログレスバー
tqdm>=4.65.0

# ログ
tensorboard>=2.13.0
wandb>=0.15.0

# ユーティリティ
click>=8.1.0
colorama>=0.4.6

# GUI (オプション)
# Tkinter は標準ライブラリ
# PyQt5>=5.15.9  # 必要に応じてコメント解除
